third_party/blink/web_tests/fast/forms/calendar-picker/week-picker-touch-operations-expected.txt

Tests if calendar picker mouse operations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
Open the month popup.
PASS internals.pagePopupWindow.global.picker.monthPopupView.isVisible is false
PASS internals.pagePopupWindow.global.picker.monthPopupView.isVisible is true
Check that touch swiping scrolls the list in the month popup.
PASS checkYearListViewScrollOffset() > 0 is true
PASS checkYearListViewScrollOffset() < 0 is true
PASS successfullyParsed is true

TEST COMPLETE
